Move errors to pages that need to render them.
