+++ src/views/SystemIntake/RequestDetails/index.tsx
@@ -24,6 +24,7 @@ import {
   clearSystemIntake
 } from 'types/routines';
 import AutoSave from 'components/shared/AutoSave';
+import { ErrorAlert, ErrorAlertMessage } from 'components/shared/ErrorAlert';
 
 type RequestDetailsProps = {
   formikProps: FormikProps<SystemIntakeForm>;
@@ -54,6 +55,30 @@ const RequestDetails = ({
 
   return (
     <>
+      {Object.keys(formikProps.errors).length > 0 && (
+        <ErrorAlert
+          classNames="margin-top-3"
+          heading="Please check and fix the following"
+        >
+          {Object.keys(flatErrors).map(key => {
+            return (
+              <ErrorAlertMessage
+                key={`Error.${key}`}
+                message={flatErrors[key]}
+                onClick={() => {
+                  const field = document.querySelector(
+                    `[data-scroll="${key}"]`
+                  );
+
+                  if (field) {
+                    field.scrollIntoView();
+                  }
+                }}
+              />
+            );
+          })}
+        </ErrorAlert>
+      )}
       <h1 className="font-heading-xl margin-top-4">Request details</h1>
       <p className="line-height-body-6">
         Provide a detailed explanation of the business need/issue/problem that
